fix(lv8): JPQL을 이용하여 일정 조회 페이징의 commentCount 계산 간 N+1 문제 개선

+++ src/main/java/io/github/seonrizee/scheduler/dto/response/SchedulePageResponse.java
@@ -11,7 +11,7 @@ public class SchedulePageResponse {
     private final Long id;
     private final String summary;
     private final String description;
-    private final Integer commentCount;
+    private final Long commentCount;
     private final Long userId;
     private final String username;
     private final LocalDateTime createdAt;
